update courseService to validate instructor data

+++ src/main/kotlin/com/kotlinspring/repository/InstructorRepository.kt
@@ -3,5 +3,5 @@ package com.kotlinspring.repository
 import com.kotlinspring.entity.Instructor
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface InstructorRepository : JpaRepository<Instructor, Long> {
+interface InstructorRepository : JpaRepository<Instructor, Int> {
 }
